Fix invalid reads
Don't use a guint16* when getting a guint property via g_object_get()!

Bug #628323.

+++ gio/tests/network-address.c
@@ -4,7 +4,7 @@ static void
 test_basic (void)
 {
   GNetworkAddress *address;
-  guint16 port;
+  guint port;
   gchar *hostname;
 
   address = (GNetworkAddress*)g_network_address_new ("www.gnome.org", 8080);
